task: #3555534 Since symfony/validator 7.4: Support for evaluating options in the base Constraint class is deprecated. Initialize properties in the constructor instead.
By: longwave
By: godotislate
By: larowlan

+++ core/modules/file/src/Plugin/Validation/Constraint/FileValidationConstraintValidator.php
@@ -48,7 +48,7 @@ class FileValidationConstraintValidator extends ConstraintValidator implements C
     // Always respect the configured maximum file size.
     $field_settings = $value->getFieldDefinition()->getSettings();
     if (array_key_exists('max_filesize', $field_settings)) {
-      $validators['FileSizeLimit'] = ['fileLimit' => Bytes::toNumber($field_settings['max_filesize'])];
+      $validators['FileSizeLimit'] = ['fileLimit' => (int) Bytes::toNumber($field_settings['max_filesize'])];
     }
     else {
       // Do not validate the file size if it is not set explicitly.
